log: 删除writer_impl文件，将其内容整合入writer中，doc: 同步文档 (#14)
log: 删除writer_impl文件，将其内容整合入writer中，doc: 同步文档 (#14)

ISSUE: --issue=#12
