+++ log/writer_test.go
